[filetree] rename "sift" folder into "popsift" and move "common" into "popsift"
With this new file tree, we update:
* cmake headers installation rules
* includes
