+++ src/popsift/assist.cu
