+++ src/popsift/assist.h
