+++ src/popsift/clamp.h
