+++ src/popsift/common/debug_macros.cpp
