+++ src/popsift/common/device_prop.cpp
