+++ src/popsift/common/device_prop.h
