+++ src/popsift/gauss_filter.h
