+++ src/popsift/s_solve.h
