+++ src/popsift/sift_desc.cu
